re-organized more components into folders
