+++ www/src/components/ecosystem/visually-hidden.js
@@ -1,5 +1,5 @@
 import React from "react"
-import { visuallyHidden } from "../utils/styles"
+import { visuallyHidden } from "../../utils/styles"
 
 const VisuallyHidden = props => {
   return <span css={visuallyHidden} {...props} />
