+++ www/src/components/layout/navigation.js
@@ -11,7 +11,7 @@ import logo from "../../assets/logo.svg"
 import logoInverted from "../../assets/logo-inverted.svg"
 import { mediaQueries } from "gatsby-design-tokens/dist/theme-gatsbyjs-org"
 import { breakpointGutter } from "../../utils/styles"
-import DarkModeToggle from "../dark-mode-toggle"
+import DarkModeToggle from "../ecosystem/dark-mode-toggle"
 
 // what we need to nudge down the navItems to sit
 // on the baseline of the logo's wordmark
