+++ www/src/pages/features.js
@@ -6,7 +6,7 @@ import { Helmet } from "react-helmet"
 
 import Button from "../components/button"
 import Layout from "../components/layout"
-import Container from "../components/container"
+import Container from "../components/ecosystem/container"
 import FooterLinks from "../components/layout/footer-links"
 import LegendTable from "../components/features/legend-table"
 import FeaturesFooter from "../components/features/features-footer"
