+++ www/src/pages/features/cms.js
@@ -7,7 +7,7 @@ import Layout from "../../components/layout"
 import PageWithSidebar from "../../components/docs/page-with-sidebar"
 import EvaluationTable from "../../components/features/evaluation-table"
 import { getFeaturesData } from "../../utils/get-csv-features-data"
-import Container from "../../components/container"
+import Container from "../../components/ecosystem/container"
 import FooterLinks from "../../components/layout/footer-links"
 import FeaturesFooter from "../../components/features/features-footer"
 import LegendTable from "../../components/features/legend-table"
