+++ www/src/pages/features/jamstack.js
@@ -6,7 +6,7 @@ import Button from "../../components/button"
 import Layout from "../../components/layout"
 import EvaluationTable from "../../components/features/evaluation-table"
 import { getFeaturesData } from "../../utils/get-csv-features-data"
-import Container from "../../components/container"
+import Container from "../../components/ecosystem/container"
 import FooterLinks from "../../components/layout/footer-links"
 import FeaturesFooter from "../../components/features/features-footer"
 import LegendTable from "../../components/features/legend-table"
