+++ www/src/templates/template-docs-markdown.js
@@ -13,7 +13,7 @@ import DocSearchContent from "../components/docsearch-content"
 import TableOfContents from "../components/docs/docs-table-of-contents"
 import FooterLinks from "../components/layout/footer-links"
 import Breadcrumb from "../components/docs/docs-breadcrumb"
-import Container from "../components/container"
+import Container from "../components/ecosystem/container"
 import PrevAndNext from "../components/prev-and-next"
 
 const containerStyles = {
