+++ www/src/templates/template-feature-comparison.js
@@ -6,7 +6,7 @@ import { css } from "@emotion/core"
 import Layout from "../components/layout"
 import PageWithSidebar from "../components/docs/page-with-sidebar"
 import FooterLinks from "../components/layout/footer-links"
-import Container from "../components/container"
+import Container from "../components/ecosystem/container"
 import EvaluationTable from "../components/features/evaluation-table"
 import LogoDictionary from "../components/features/logo-dictionary"
 import LegendTable from "../components/features/legend-table"
